Move Main Tree Logic to ScrapPackedLibrary for reuse in tree command
